+++ test/test_jbpy.py
@@ -30,6 +30,7 @@ def test_roundtrip_jitc_quicklook(filename, tmp_path):
 
 EXPECTED_TRES = (
     "BLOCKA",
+    "EXOPTA",
     "SECTGA",
     "STDIDC",
     "USE00A",