+++ test/test_jbpy.py
@@ -31,6 +31,7 @@ def test_roundtrip_jitc_quicklook(filename, tmp_path):
 EXPECTED_TRES = (
     "BLOCKA",
     "GEOPSB",
+    "ICHIPB",
     "SECTGA",
     "STDIDC",
     "USE00A",