+++ jbpy/extensions/tre/__init__.py
@@ -1,6 +1,7 @@
-from ._core import EncodedFixedPoint, FloatFormat
+from ._core import EncodedFixedPoint, FlexibleFloat, FloatFormat
 
 __all__ = [
     "EncodedFixedPoint",
+    "FlexibleFloat",
     "FloatFormat",
 ]