+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Encoded range checking to `core.Field`
-- TRE support: BLOCKA, USE00A, STDIDC
+- TRE support: BLOCKA, EXOPTA, USE00A, STDIDC
 - `FloatFormat` converter
 
 ### Changed