+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Encoded range checking to `core.Field`
-- TRE support: BLOCKA, GEOPSB, USE00A, STDIDC
+- TRE support: BLOCKA, GEOPSB, ICHIPB, STDIDC, USE00A
 - `FloatFormat` converter
 
 ### Changed