+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.4.0] - 2025-11-05
+
 ### Added
 - Encoded range checking to `core.Field`
 - TRE support: BLOCKA, EXOPTA, GEOPSB, ICHIPB, J2KLRA, PRJPSB, REGPTB, RPC00B, STDIDC, USE00A
@@ -66,7 +69,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic JBP functionality copied from SARkit's `_nitf_io.py`
 - TRE support: SECTGA
 
-[unreleased]: https://github.com/ValkyrieSystems/jbpy/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/ValkyrieSystems/jbpy/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/ValkyrieSystems/jbpy/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/ValkyrieSystems/jbpy/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/ValkyrieSystems/jbpy/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ValkyrieSystems/jbpy/releases/tag/v0.1.0